Support for multiline doctests: #50
There is already some `filterExpression' in Interpreter.hs, but it
isn't clear that such a function ever succeeded with :{ input, since
the parser previously would have each line as a separate Example.

This change to the parser sends all the content between :{ :} to ghci at
once. This seems to be necessary to support comments like:

    >>> :{
    >>> let
    >>>   x = 1
    >>>   y = 2
    >>>  in x + y
    >>> :}
    3