Avoid QuickCheck for properties without free variables
Properties that have no free variables and have type `Bool` can be checked
to be `True` without using QuickCheck. Properties with free variables need
QuickCheck. Some properties may not have free variables but if the type
is not `Bool` we use QuickCheck on the assumption that they are of some
other `Test.QuickCheck.Testable` instance.